pyproject.toml
[tool.poetry]
name = "cloudflare_dyndns"
version = "0.1.0"
description = "Script to set IP address on Cloudflare for dynamic host IP"
authors = ["William Hayes <william.s.hayes@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.22.0"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"


[tool.black]
line-length = 100
target-version = ['py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
[tool.isort]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=false
use_parentheses=true
line_length=100
import_heading_stdlib="Standard Library"
import_heading_thirdparty="Third Party Imports"
import_heading_firstparty="Local Imports"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
